Rearrange to get access to the channel setting
We actually do not want to attempt to access the channel at all.
That would cause quite a conundrum and interfere with the way it works.
What we do want is to access the variable that was set from the channel.
You will see this line in authz.go to do that:

	version := s.vSwitch.version

But that means we need `vSwitch` on the server, which needed to modify
the fields of authzServer to provide it (instead of versionGetter).
And to do that we need to move the `versionSwitch` type and its paraphernalia
from server.go to authz.go (primarily to avoid circular references).
